+++ python/sdssdb/peewee/lvmdb/lvmopsdb.py
@@ -105,6 +105,16 @@ class Observation(LVMOpsBase):
         table_name = 'observation'
 
 
+class Redo(LVMOpsBase):
+    tile = ForeignKeyField(column_name='tile_id',
+                           field='tile_id',
+                           model=Tile)
+    nexp = IntegerField(default=1)
+
+    class Meta:
+        table_name = 'redo'
+
+
 class Weather(LVMOpsBase):
     pk = AutoField()
     observation = ForeignKeyField(column_name='obs_id',