+++ schema/sdss5db/catalogdb/mwsall/mwsall_fibermap_pkey.sql
@@ -1,0 +1,4 @@
+\o mwsall_fibermap_pkey.out
+-- mwsall_fibermap column targetid has duplicates.
+-- alter table catalogdb.mwsall_fibermap add primary key(targetid);
+\o